flask/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
